Add duration Trait
