+++ src/Illuminate/Cache/composer.json
@@ -9,8 +9,7 @@
     ],
     "require": {
         "php": ">=5.4.0",
-        "illuminate/support": "4.2.*",
-        "nesbot/carbon": "~1.0"
+        "illuminate/support": "4.2.*"
     },
     "require-dev": {
         "illuminate/database": "4.2.*",
