+++ src/Illuminate/Queue/composer.json
@@ -13,8 +13,7 @@
         "illuminate/container": "4.2.*",
         "illuminate/http": "4.2.*",
         "illuminate/support": "4.2.*",
-        "symfony/process": "2.5.*",
-        "nesbot/carbon": "~1.0"
+        "symfony/process": "2.5.*"
     },
     "require-dev": {
         "illuminate/cache": "4.2.*",
