+++ src/Illuminate/Support/composer.json
@@ -12,7 +12,8 @@
     },
     "require-dev": {
         "jeremeamia/superclosure": "~1.0",
-        "patchwork/utf8": "1.1.*"
+        "patchwork/utf8": "1.1.*",
+        "nesbot/carbon": "~1.0"
     },
     "autoload": {
         "psr-0": {
